Agent query: Please verify if the sales chart is now loading correctly without any GROUP BY clause errors. Can you check if the data points are showing up properly for different time periods (daily/weekly/monthly/yearly)?

+++ server/routes.ts
@@ -126,8 +126,8 @@ export function registerRoutes(app: Express) {
       const salesData = await db
         .select({
           date: sql`date_trunc(${period}, ${sales.createdAt})::date`,
-          total: sql`COALESCE(SUM(${sales.total}), 0)::decimal`,
-          count: sql`COUNT(*)::integer`
+          total: sql`COALESCE(SUM(${sales.total}), 0)`,
+          count: sql`COUNT(*)`
         })
         .from(sales)
         .where(sql`${sales.createdAt} >= ${startDate}`)
